refactor: share parts of the workflow between amor and estia
