+++ src/ess/amor/normalization.py
@@ -8,6 +8,7 @@ from ..reflectometry.supermirror import (
     supermirror_reflectivity,
 )
 from ..reflectometry.types import (
+    CoordTransformationGraph,
     DetectorSpatialResolution,
     ReducedReference,
     ReducibleData,
@@ -22,7 +23,6 @@ from .resolution import (
     sample_size_resolution,
     wavelength_resolution,
 )
-from .types import CoordTransformationGraph
 
 
 def mask_events_where_supermirror_does_not_cover(
