+++ src/ess/amor/types.py
@@ -9,8 +9,6 @@ WavelengthResolution = NewType("WavelengthResolution", sc.Variable)
 AngularResolution = NewType("AngularResolution", sc.Variable)
 SampleSizeResolution = NewType("SampleSizeResolution", sc.Variable)
 
-CoordTransformationGraph = NewType("CoordTransformationGraph", dict)
-
 
 class ChopperFrequency(sciline.Scope[RunType, sc.Variable], sc.Variable):
     """Frequency of the choppers used in the run."""
